requirements.txt
streamlit==1.32.0
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.1
numpy==1.26.4
plotly==5.18.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
httpx==0.27.0
pytest==8.0.0
black==24.2.0
flake8==7.0.0 